+++ tycho-integration-test/src/stream_processor/protocol_stream_processor.rs
@@ -16,10 +16,7 @@ use tycho_simulation::{
         protocol::{
             aerodrome_slipstreams::state::AerodromeSlipstreamsState,
             ekubo::state::EkuboState,
-            filters::{
-                balancer_v2_pool_filter, curve_pool_filter, fluid_v1_paused_pools_filter,
-                uniswap_v4_euler_hook_pool_filter,
-            },
+            filters::{balancer_v2_pool_filter, fluid_v1_paused_pools_filter},
             fluid::FluidV1,
             pancakeswap_v2::state::PancakeswapV2State,
             uniswap_v2::state::UniswapV2State,
@@ -132,16 +129,8 @@ impl ProtocolStreamProcessor {
                     )
                     .exchange::<UniswapV4State>("uniswap_v4", tvl_filter.clone(), None)
                     .exchange::<EkuboState>("ekubo_v2", tvl_filter.clone(), None)
-                    .exchange::<EVMPoolState<PreCachedDB>>(
-                        "vm:curve",
-                        tvl_filter.clone(),
-                        Some(curve_pool_filter),
-                    )
-                    .exchange::<UniswapV4State>(
-                        "uniswap_v4_hooks",
-                        tvl_filter.clone(),
-                        Some(uniswap_v4_euler_hook_pool_filter),
-                    )
+                    .exchange::<EVMPoolState<PreCachedDB>>("vm:curve", tvl_filter.clone(), None)
+                    .exchange::<UniswapV4State>("uniswap_v4_hooks", tvl_filter.clone(), None)
                     .exchange::<EVMPoolState<PreCachedDB>>(
                         "vm:maverick_v2",
                         tvl_filter.clone(),